Bugfix/spline test data generator limited ranges (#1040)
* spline #1017 init generator enhancements

* spline #1017 lincenses

* spline #1017 graph type and variable parameters

* Delete lineage-4ops-2reads-0attr-3expr-event.json

* Delete lineage-4ops-2reads-0attr-3expr-plan.json

* spline #1017 feedback

* spline #1017 expressions per each attribute

* spline #1017 done with testing

* spline #1017 license

* spline #1017 final modifications, tested with running service

* spline #1017 code feedback

* spline #1017 updated script, usability, orphaned reads

* spline #1017 some feedback

* spline #1017 final touches

* spline #1039 fixes bug

+++ test-data-generator/src/main/scala/za/co/absa/spline/testdatagen/NumericValue.scala
@@ -43,7 +43,7 @@ object NumericValue {
       case ConstantPattern(c) if c.toInt > 0 => Constant(c.toInt)
       case ConstantPattern(c) => throw new IllegalArgumentException(s"Invalid name $c, number should be positive")
       case VariablePattern(start: String, end: String, step: String)
-        if start.toInt > 0 && end.toInt > 0 && step.toInt > 0 && end > start =>
+        if start.toInt > 0 && end.toInt > 0 && step.toInt > 0 && end.toInt > start.toInt =>
         Variable(start.toInt, end.toInt, step.toInt)
       case VariablePattern(_, _, _) =>
         throw new IllegalArgumentException(s"Invalid range $param, pattern should be start-end/step, each one > 0")
